refactor(scripts): split backfill_history + backfill_intelligence into facade + mixins
- backfill_history.py (677L→~250L): facade + 3 mixins (DataFetchMixin, ProcessingMixin, PersistenceMixin)
- backfill_intelligence.py (658L→~150L): facade + 2 mixins (DataFetchMixin, AnalysisMixin)

Follows established pattern from PR #178 (weekly_aggregator, insert_analysis) and PR #181 (zwift_scraper).
